Move contributor patterns to a package'
